\name{INFCstats}
\alias{INFCstats}
\docType{data}
\title{Equation statistics}
\description{A \code{data.frame} containing values for variance and covariance matrices}
\format{A data frame with 220 observations on the following 21 variables.}
\details{
The list of species and of their \code{spg}-codes is:\cr \tabular{ll}{Abies alba:\tab \code{Abal}\cr Acer campestre:\tab \code{Acca}\cr Acer monspessolanum:\tab \code{Acmo}\cr Acer opalus:\tab \code{Acop}\cr Acer pseudoplatanus:\tab \code{Acps}\cr Alnus cordata:\tab \code{Alco}\cr Alnus
glutinosa:\tab \code{Algl}\cr Carpinus betulus:\tab \code{Cabe}\cr Castanea sativa:\tab \code{Casa}\cr Cupressus spp:\tab \code{Cusp}\cr Eucalyptus occidentalis:\tab \code{Euoc}\cr Fagus sylvatica:\tab \code{Fasy}\cr Fraxinus
angustifolia:\tab \code{Fran}\cr Fraxinus excelsior:\tab \code{Frex}\cr Fraxinus ornus:\tab \code{Fror}\cr Laburnum alpinum:\tab \code{Laal}\cr Larix decidua:\tab \code{Lade}\cr Ostrya carpinifolia:\tab \code{Ossp}\cr Picea abies:\tab \code{Piab}\cr Pinus cembra:\tab \code{Pice}\cr Pinus halepensis:\tab \code{Piha}\cr Pinus nigra var. laricio:\tab \code{Pila}\cr
Pinus nigra var. nigra:\tab \code{Pini}\cr Pinus pinaster:\tab
\code{Pips}\cr Pinus pinea:\tab \code{Pipi}\cr Pinus radiata:\tab
\code{Pira}\cr Pinus strobus:\tab \code{Pist}\cr Pinus sylvestris:\tab
\code{Pisy}\cr Populus canescens:\tab \code{Poca}\cr Populus nigra:\tab
\code{Poni}\cr Populus tremula:\tab \code{Potr}\cr Prunus avium:\tab
\code{Prav}\cr Pseudotsuga menziesii:\tab \code{Psme}\cr Quercus
cerris:\tab \code{Quce}\cr Quercus ilex:\tab \code{Quil}\cr Quercus
pubescens:\tab \code{Qupu}\cr Robinia pseudoacacia:\tab \code{Rops}\cr
Salix alba:\tab \code{Saal}\cr Salix caprea:\tab \code{Saca}\cr Sorbus
aria:\tab \code{Soar}\cr Tilia cordata:\tab \code{Tico}\cr Tilia
platyphyllos:\tab \code{Tipl}\cr Ulmus minor:\tab \code{Ulmi}\cr }
}
\references{
Tabacchi G., Di Cosmo L., Gasparini P., Morelli S., 2011a.
\emph{Stima del volume e della fitomassa delle principali specie forestali
italiane. Equazioni di previsione, tavole del volume e tavole della
fitomassa arborea epigea. Stima del volume e della fitomassa delle
principali specie forestali italiane. Equazioni di previsione, tavole del
volume e tavole della fitomassa arborea epigea}. 412 pp. [ITA, ita]
}
\keyword{datasets}