\name{mydata}
\alias{mydata}
\docType{data}
\title{
Tree list dataset
}
\description{
An example of possible dataset.
}
\usage{data("mydata")}
\format{
  A data frame with 207 observations on the following 15 variables.
  \describe{
    \item{\code{sito}}{a numeric vector}
    \item{\code{compartimento}}{a numeric vector}
    \item{\code{plot}}{a numeric vector}
    \item{\code{tree_id}}{a numeric vector}
    \item{\code{specie}}{a factor with levels \code{Acer camp} \code{Carp bet} \code{Cerro} \code{Ciliegio selv} \code{Cornus mas} \code{Melo selv} \code{Nocciolo} \code{Olmo spp} \code{Orniello} \code{Q robur} \code{Sorbo}}
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{nrad}}{a numeric vector}
    \item{\code{srad}}{a numeric vector}
    \item{\code{erad}}{a numeric vector}
    \item{\code{wrad}}{a numeric vector}
    \item{\code{altezza}}{a numeric vector}
    \item{\code{pos}}{a numeric vector}
    \item{\code{diam}}{a numeric vector}
    \item{\code{status}}{a numeric vector}
  }
}
\details{
possibili valori
}
\references{
Puletti et al 2017
}
\examples{
data(mydata)
## maybe str(mydata) ; plot(mydata) ...
}
\keyword{datasets}
